Rename WebSocket.ExecutionMode.SIMULTANEOUS to CONCURRENT
